requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14.0
pillow>=9.0.0  # Para manipulação de imagens (ImageField)
python-decouple>=3.8  # Para gerenciar variáveis de ambiente
markdown>=3.5  # Para formatação de saída da API em navegadores
django-filter>=23.2  # Para filtrar resultados da API